Allow apps to return errors for events.
Apps can either throw an error if they want to respond synchronously
to events, or pass an error object to the callback as the first param
for asynchronous responses.

Note that this breaks the events API for apps that use async event
handling. The first param is now an error and not the body.